Ensure we get an integer

+++ src/Pager/BasePager.php
@@ -129,7 +129,7 @@ abstract class BasePager implements \Serializable, PagerInterface
             $nbLinks = $this->getMaxPageLinks();
         }
         $links = [];
-        $tmp = $this->page - floor($nbLinks / 2);
+        $tmp = $this->page - (int) floor($nbLinks / 2);
         $check = $this->lastPage - $nbLinks + 1;
         $limit = $check > 0 ? $check : 1;
         $begin = $tmp > 0 ? ($tmp > $limit ? $limit : $tmp) : 1;
